Rewrite impute condition evaluation and how data is cached during Jinja2 template rendering and expression evaluation
Work in Progress! Page load times are reduced, but it might be because something isn't happening that should be happening. Some impute conditions may be coming up with the wrong values or something. It might also be a good idea to split this into three commits.

This commit began to address the speed issue in generating the question pages. In complex compliance apps with a very large number of questions, page load times were reaching tens of seconds.

This commit makes three sets of related changes:

One of the core subroutines in the GovReady-Q code is evaluate_module_state, which computes two things simultaneously:

* The imputed answer to each question for which an impute condition applies. Because impute conditions rely on the answers to other questions, dependencies must be figured out before the questions that depend on them.
* A list of questions the user may answer next. We don't know whether a question should be shown or not until all of its impute conditions can be evaluated, and we can't show question prompts and other template text until the questions that are used in the template have answers themselves. The list of questions that may be answered contains those questions that are not yet answered and whose dependencies all have answers.

When we first developed Q, we asked questions in an order determined entirely by their dependencies. We started with the output documents and searched for questions used in the document, and then worked backwards through dependencies until we found a question with no unanswered dependencies, and that's what the user was asked next. We later found that it was important to give users a linear context to understand their progress through a module, and we found that we wanted all questions in a module to be answered (or imputed) even if they were not used in an output document, so the original logic has pretty much been gone, but the subroutine was still written to follow question dependencies recursively.

This commit re-writes evaluate_module_state so that it is not recursive. It now proceeds through the questions in order. For each question, if it depends on a previous question that has not yet been answered, it remains unanswered. Otherwise, if it has an imput condition that evaluates to true, the imputed value is set as the question's answer. Otherwise, if the user has answered the question, that answer is used. The function is much simpler now. This change has no impact on speed but makes the code more maintainable.

guidedmodules/module_logic.py:

* `evaluate_module_state` is rewritten. It no longer uses walk_module_questions, and it was the last method calling walk_module_questions, so walk_module_questions is deleted.

When we evaluate impute conditions (as described above) and when we generate output documents, we evaluate Jinja2 expressions and render Jinja2 templates. The expressions and templates can use/include the answers to questions, e.g. the answers to questions can appear as `{{question_id}}`. The set of variables available during expression evaluation and template rendering is provided by a TemplateContext instance (which is a class we created that implements a mapping (dict-like) interface).

The TemplateContext class constructor was very complicated with a lot of parameters. It's now simplified by grouping all of those options into a single TemplateRenderingOptions class instance, a new class created in this commit.

guidedmodules/module_logic.py:

* TemplateRenderingOptions class added.
* All new instances of TemplateContext now create and pass a TemplateRenderingOptions instance rather than passing a lot of parameters.
* Titles for tasks and projects can be computed dynamically by rendering a Jinja2 template. The is_computing_title option is changed from being a boolean to being a Task instance, the Task whose title is being computed. When computing this generated title, we have to disable accessing the title from within the expression/template, otherwise we would have infinite recursion.

There is a complicated range of ways that project data can be accessed through Jinja2 expressions and template variables. In addition to simply `{{question_id}}`, there can also be recursive accesses like `{{question_id.inner_question_id}}` (when the outer question is a module-type question), or the entire compliance app's answers can be accessed through a `project` variable, e.g. `{{project.module.question_id}}`.

My hypothesis for why page loads were so slow was redundant database queries or redundantly calling evaluate_module_state multiple times for the same module. For instance, every time `{{project....}}` appears, the result of evaluate_module_state on the root task of the project is needed. `{{project....}}` can appear many times in an expression and template, but it should not need to go to the database or evaluate_module_state more than once.

This commit implements a new short-term caching mechanism so that for the duration of a page request, module metadata and question answers are held in memory after they are first requested so that they don't need to be accessed again. A new cache "TemplateEvaluationDataCache" object is passed down recursively through the expression evaluation and template rendering routines. Since these routines don't modify any question answers, we can assume that any data requested from the database remains unchanged for the duration of these calls.

To make this change, I refactored some of the

guidedmodules/models.py:

* Task class: Titles for tasks and projects can be computed dynamically by rendering a Jinja2 template. This was done in a property. It is now done in a method that takes an optional cache object so that it can make use of the new cache mechanism. render_output_documents also gets a new cache parameter.
* render_content gets a new cache parameter so it can take advantage of the new cache mechanism.

guidedmodules/module_logic.py:

* Some caching in TemplateContext and RenderedAnswer is removed.
* RenderedAnswer.parent_context is renamed to template_context.

guidedmodules/tests.py:

* Tests updated.

guidedmodules/views.py:

* In order to be able to pass a single TemplateEvaluationDataCache object to all of the rendering that happens in the question-related views, the call to task.get_answers().with_extended_info() is moved from the task_view decorator into the particualar views that need it (next_question, show_question, task_finished, download_module_output), and when it is not the only rendering that occurs in the view (show_question, task_finished), a TemplateEvaluationDataCache is passed to it and the other template rendering.