Adjustment for data_cache
The idea is that we have to do a lot of operations / database queries during the rendering of a page and these often involve querying the same data. So I created a class TemplateEvaluationDataCache which is meant to serve as a cache for the duration of a page load, and an instance of this class is passed around during template rendering etc etc.

(At the end of the request, the object is reclaimed by Python automatically.)

It looks like I didn't finish making the relevant changes in the tests. The way I figured this out was to look at how TemplateContext was instantiated elsewhere.  The third argument is sometimes passed TemplateEvaluationDataCache() , a new instance of a cache object.

So, ideally, there would be one instance of TemplateEvaluationDataCache created per request. It would be created in the view and passed down where needed. But in tests, where there is no request, it would make sense to just create a new instance of TemplateEvaluationDataCache for the purposes of the test. (edited)

And because the branch isn't entirely complete, it's probably not actually creating a new TemplateEvaluationDataCache in each request - it's probably creating a few as needed. Ideally the only place a TemplateEvaluationDataCache would be created without be in a few function (or test), but for now it's creating it a little deeper in the routines.

The tests are erroring b/c the TemplateEvaluationDataCache is not being created in tests.py?

Just add TemplateEvaluationDataCache() as the missing third argument.